test/requirements.txt
# Python dependencies for the full project (Node packages intentionally excluded)
# Standard-library modules are not listed.

python-dotenv>=1.0.0,<2.0.0
langgraph>=0.0.1
langsmith>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
